move some files and introduce UnitsJsonExporter
