+++ tools/scdat2json/to_json.cpp
@@ -1,0 +1,117 @@
+#include "to_json.h"
+
+using namespace std;
+using namespace dat;
+
+void to_json(json &j, units_dat_t::hit_points_type_t* t)
+{
+  j = json{ {"hitpoints", t->hitpoints()} };
+}
+
+void to_json(json &j, units_dat_t::special_ability_flags_type_t* t)
+{
+  j = json{
+    {"building", t->building()},
+    {"addon", t->addon()},
+    {"flyer", t->flyer()},
+    {"resourceminer", t->resourceminer()},
+    {"subunit", t->subunit()},
+    {"flyingbuilding", t->flyingbuilding()},
+    {"hero", t->hero()},
+    {"regenerate", t->regenerate()},
+    {"animatedidle", t->animatedidle()},
+    {"cloakable", t->cloakable()},
+    {"twounitsinoneegg", t->twounitsinoneegg()},
+    {"singleentity", t->singleentity()},
+    {"resourcedepot", t->resourcedepot()},
+    {"resourcecontainter", t->resourcecontainter()},
+    {"robotic", t->robotic()},
+    {"detector", t->detector()},
+    {"organic", t->organic()},
+    {"requirescreep", t->requirescreep()},
+    {"unused", t->unused()},
+    {"requirespsi", t->requirespsi()},
+    {"burrowable", t->burrowable()},
+    {"spellcaster", t->spellcaster()},
+    {"permanentcloak", t->permanentcloak()},
+    {"pickupitem", t->pickupitem()},
+    {"ignoresupplycheck", t->ignoresupplycheck()},
+    {"usemediumoverlays", t->usemediumoverlays()},
+    {"uselargeoverlays", t->uselargeoverlays()},
+    {"battlereactions", t->battlereactions()},
+    {"fullautoattack", t->fullautoattack()},
+    {"invincible", t->invincible()},
+    {"mechanical", t->mechanical()},
+    {"producesunits", t->producesunits()}
+  };
+}
+
+void to_json(json &j, units_dat_t::staredit_placement_box_type_t* t)
+{
+  j = json{ {"width", t->width()}, {"height", t->height()} };
+}
+
+void to_json(json &j, units_dat_t::addon_position_type_t* t)
+{
+  j = json{ {"horizontal", t->horizontal()}, {"vertical", t->vertical()} };
+}
+
+void to_json(json &j, units_dat_t::unit_dimension_type_t* t)
+{
+  j = json{ {"left", t->left()}, {"up", t->up()}, {"right", t->right()}, {"down", t->down()} };
+}
+
+void to_json(json &j, units_dat_t::staredit_group_flags_type_t* t)
+{
+  j = json{
+    {"zerg", t->zerg()},
+    {"terran", t->terran()},
+    {"protoss", t->protoss()},
+    {"men", t->men()},
+    {"building", t->building()},
+    {"factory", t->factory()},
+    {"independent", t->independent()},
+    {"neutral", t->neutral()}
+  };
+}
+
+void to_json(json &j, units_dat_t::staredit_availability_flags_type_t* t)
+{
+  j = json{
+    {"non_neutral", t->non_neutral()},
+    {"unit_listing", t->unit_listing()},
+    {"mission_briefing", t->mission_briefing()},
+    {"player_settings", t->player_settings()},
+    {"all_races", t->all_races()},
+    {"set_doodad_state", t->set_doodad_state()},
+    {"non_location_triggers", t->non_location_triggers()},
+    {"unit_hero_settings", t->unit_hero_settings()},
+    {"location_triggers", t->location_triggers()},
+    {"brood_war_only", t->brood_war_only()}
+  };
+}
+
+namespace dat {
+
+void to_json(json &j, TblEntry t)
+{
+  j = json{
+    {"name1", t.name1},
+    {"name2", t.name2},
+    {"name3", t.name3},
+    {"shortcut_pos", t.shortcut_pos},
+    {"shortcut", t.shortcut}
+  };
+}
+
+void to_json(json &j, Unit *u)
+{
+  j = json{
+    {"id", u->createID()},
+    {"hitpoints", u->hitpoints()},
+    {"name", u->name()}
+
+  };
+}
+
+}
