+++ tools/scdat2json/to_json.h
@@ -1,0 +1,34 @@
+/*
+ * to_json.h
+ *
+ *      Author: Andreas Volz
+ */
+
+#ifndef TO_JSON_H
+#define TO_JSON_H
+
+#include "dat/DataHub.h"
+#include "dat/Unit.h"
+
+// -> units_dat.ksy
+
+void to_json(json &j, units_dat_t::hit_points_type_t* t);
+void to_json(json &j, units_dat_t::special_ability_flags_type_t* t);
+void to_json(json &j, units_dat_t::staredit_placement_box_type_t* t);
+void to_json(json &j, units_dat_t::addon_position_type_t* t);
+void to_json(json &j, units_dat_t::unit_dimension_type_t* t);
+void to_json(json &j, units_dat_t::staredit_group_flags_type_t* t);
+void to_json(json &j, units_dat_t::staredit_availability_flags_type_t* t);
+
+// -> file_tbl.ksy
+
+// -> specific wrapper objects in dat namespace
+
+namespace dat {
+
+void to_json(json &j, TblEntry t);
+void to_json(json &j, Unit *u);
+
+} /* namespace dat */
+
+#endif /* TO_JSON_H_ */
